**Biass correction**
As of e-mail by david
```
 I changed the code so that after thinning or mortality,
the getTotalsAndProportions and getDiameterDistributions will only be run
if StemNo changed.

However, I also wrote the following to each instance where they are called:
Call getTotalsAndProportions
Call getDiameterDistributions(month)
Call getTotalsAndProportions
Call getDiameterDistributions(month)

This is because the bias will only be partly corrected when it is called
once. Going from the initial biased values to completely unbiased values is
probably asymptotic, such that the first time it is run will correct
perhaps more than half of the bias, and then the second time will correct a
lot of the remaining bias – although to correct for 100% of the bias would
require them to be called many times, with each subsequent run correcting a
much smaller proportion of bias. I think this is okay to do this because
the main use of these functions is to remove the bias in order to increase
the precision of the predictions and this circularity won’t cause the model
run to fail.
```